Prepared new util package for modules with static helper functions and renamed the render package to avoid conflicts and confusion with the render module.
