--- scripts/deforum_helpers/render_data/__init__.py
@@ -1,4 +1,0 @@
-from .schedule import Schedule
-
-# initialization is initialized last because it refs to the other data objects.
-from .initialization import RenderInit, RenderInitArgs
