+++ scripts/deforum_helpers/rendering/data/schedule.py
@@ -1,4 +1,5 @@
 import dataclasses
+
 from typing import Optional, Any
