+++ scripts/deforum_helpers/rendering/util/__init__.py
@@ -1,0 +1,3 @@
+# All modules in this package are intended to not hold or change any state.
+# Only use static class methods or loose defs that don't change anything in the arguments passed.
+from .memory_utils import MemoryUtils
